Adds species catalogue.

+++ inc/texture_manager.hpp
@@ -15,10 +15,10 @@ namespace ld40 {
 	class TextureManager {
 		public:
 			void load_sheet(std::string sheet);
-			std::pair<const sf::Texture*, sf::Rect<float>> get_texture(std::string texture) const;
+			std::pair<const sf::Texture*, sf::Rect<int>> get_texture(std::string texture) const;
 		private:
 			std::unordered_map<std::string, std::unique_ptr<sf::Texture>> sheets;
-			std::unordered_map<std::string, std::pair<const sf::Texture*, sf::Rect<float>>> textures;
+			std::unordered_map<std::string, std::pair<const sf::Texture*, sf::Rect<int>>> textures;
 			std::experimental::filesystem::path get_full_path(std::string_view filename);
 	};
 }
